Add helper method to generate names
Generating a fresh name when the app loads will save users a click.
("Don't Make Me Click" right?) Since we will be using the name
generating functionality in two places, it makes sense to pull it out
into a helper method.

This seems like it might work better as a method of the stores.