Make changes to use flux when rendering current name
There was a draft of the actions, stores, and dispatcher written, but
none of them were really being used by the view (the components). This
commit wires it all up:
 * The entry point kicks off a helper function that simulates fetching
the first "currentName" value from the API and into the stores.
 * The component gets the "currentName" from the stores and listens for
changes to the stores.
 * Clicking the button triggers a view action that in turn triggers the
server action which fetches a new name from the API. (Not sure that this
is a good use of "server" actions - should those just be updates pushed
from the server with something like websockets?)
 * The new "currentName" triggers a change event in the stores, which
triggers an update in the view.

Next steps:
 * Separate out the logic for handling a list of saved names from the
logic for tracking the "currentName" (should probably be a separate set
of actions and stores)
 * Create the components to CRUD some names that are persisted to
localStorage