Add server actions and API utils for mocking CRUD of names
To use flex for a more full CRUD of a resource, I'm creating a
'namesList' for saving favorite names. This commit adds API helpers to
persist creating and deleting names on a mock server (localStorage) and
a set of server actions to serve as callbacks for successful persistence
to the server.