Namespace the constants
We are using a set of global constants to hold the keywords used for
CRUD events. This is a good thing, because with the introduction of a
new resource it won't do to say 'create' without any mention of what is
being created. This commit namespaces the constants and puts them in
sync with the new server actions.