Fetch names from server in NameApp component
To actually start using the list of names, we need to include it in the
state of the NameApp component. Doing so brought up the issue that I had
forgotten to JSON.stringify and JSON.parse the data for the namesList
when taking it in and out of localStorage, so this fixes that and a
couple of other minor issues.

I also ran webpack to regenerate the bundle.